% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{br_economy_model}
\alias{br_economy_model}
\title{Macroeconomic Data for Brazil}
\format{
A data frame with 76 rows and 9 variables
}
\source{
Brazilian Central Bank and Author's calculations.
}
\usage{
br_economy_model
}
\description{
The dataset contains quarterly macroeconomic data for the Brazilian economy.
}
\examples{
data(br_economy_model)
head(br_economy_model)
}
\keyword{datasets}
