% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{brl_usd}
\alias{brl_usd}
\title{BRL/USD Exchange Rate}
\format{
A data frame with 3014 rows and 2 variables
}
\source{
Brazilian Central Bank.
}
\usage{
brl_usd
}
\description{
The dataset contains the daily exchange rate between Brazilian Real (BRL) and the US Dollar (USD)
}
\examples{
data(brl_usd)
head(brl_usd)
}
\keyword{datasets}
