% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{covid_cases}
\alias{covid_cases}
\title{Number of COVID confirmed cases and deaths}
\format{
A data frame with 23125 rows and 4 variables
}
\source{
Author
}
\usage{
covid_cases
}
\description{
This dataset contains daily values of COVID confirmed cases and deaths for several countries
}
\examples{
data(covid_cases)
head(covid_cases)
}
\keyword{datasets}
