% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{cpi_br}
\alias{cpi_br}
\title{Headline CPI Data for Brazil}
\format{
A data frame with 224 rows and 2 variables
}
\source{
The Brazilian Institute of Geography and Statistics - IBGE
}
\usage{
cpi_br
}
\description{
This dataset contains monthly headline CPI data for Brazil
}
\examples{
data(cpi_br)
head(cpi_br)
}
\keyword{datasets}
