% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{cpi_item_br}
\alias{cpi_item_br}
\title{Disaggregated CPI Data for Brazil}
\format{
A data frame with 121 rows and 53 variables
}
\source{
The Brazilian Institute of Geography and Statistics - IBGE
}
\usage{
cpi_item_br
}
\description{
This dataset contains monthly Brazilian CPI data, accumulated over 12 months, disaggregated by item
}
\examples{
data(cpi_item_br)
head(cpi_item_br)
}
\keyword{datasets}
