% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{cpi_us}
\alias{cpi_us}
\title{Headline CPI Data for the US}
\format{
A data frame with 156 rows and 4 variables
}
\source{
FRED
}
\usage{
cpi_us
}
\description{
This dataset contains monthly headline CPI data for the US
}
\examples{
data(cpi_us)
head(cpi_us)
}
\keyword{datasets}
