% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{credit_card_br}
\alias{credit_card_br}
\title{Alternative data for measuring economic activity in Brazil}
\format{
A data frame with 2142 rows and 12 variables
}
\source{
Author
}
\usage{
credit_card_br
}
\description{
This dataset contains daily aggregated values of credit card transactions across a range of retail sectors.
}
\examples{
data(credit_card_br)
head(credit_card_br)
}
\keyword{datasets}
