% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{europe_population}
\alias{europe_population}
\title{Population of European Countries}
\format{
A data frame with 55 rows and 2 variables
}
\source{
Our World in Data
}
\usage{
europe_population
}
\description{
This dataset contains the population of European countries in 2022.
}
\examples{
data(europe_population)
head(europe_population)
}
\keyword{datasets}
