% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{gdp_br}
\alias{gdp_br}
\title{Brazilian GDP}
\format{
A data frame with 116 rows and 2 variables
}
\source{
The Brazilian Institute of Geography and Statistics - IBGE
}
\usage{
gdp_br
}
\description{
This dataset contains seasonally adjusted quarterly real GDP data for Brazil
}
\examples{
data(gdp_br)
head(gdp_br)
}
\keyword{datasets}
