% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{gdp_contribution_us}
\alias{gdp_contribution_us}
\title{Contributions to percent change in real GDP for the US}
\format{
A data frame with 24 rows and 5 variables
}
\source{
FRED.
}
\usage{
gdp_contribution_us
}
\description{
This dataset contains data on sectoral contributions to the percentage change in US GDP.
}
\examples{
data(gdp_contribution_us)
head(gdp_contribution_us)
}
\keyword{datasets}
