% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{gdp_ip_br}
\alias{gdp_ip_br}
\title{Brazilian GDP and Industrial Production}
\format{
A data frame with 303 rows and 3 variables
}
\source{
The Brazilian Institute of Geography and Statistics - IBGE
}
\usage{
gdp_ip_br
}
\description{
This dataset contains seasonally adjusted quarterly GDP data and seasonally adjusted monthly industrial production data for Brazil.
}
\examples{
data(gdp_ip_br)
head(gdp_ip_br)
}
\keyword{datasets}
