% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{nominal_deflator_gdp_us}
\alias{nominal_deflator_gdp_us}
\title{Yearly Nominal GDP and GDP Deflator for the US}
\format{
A data frame with 96 rows and 3 variables
}
\source{
FRED.
}
\usage{
nominal_deflator_gdp_us
}
\description{
This dataset contains yearly data on nominal GDP and GDP Deflator for the US.
}
\examples{
data(nominal_deflator_gdp_us)
head(nominal_deflator_gdp_us)
}
\keyword{datasets}
