% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{owid_covid}
\alias{owid_covid}
\title{COVID-19 Data}
\format{
A data frame with 303 rows and 3 variables
}
\source{
Our World in Data
}
\usage{
owid_covid
}
\description{
This dataset contains COVID-19 data for multiple countries.
}
\examples{
data(owid_covid)
head(owid_covid)
}
\keyword{datasets}
