% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{perfume_prices_cpi_br}
\alias{perfume_prices_cpi_br}
\title{Dataset on perfume prices in Brazil}
\format{
A data frame with 58 rows and 35 variables
}
\source{
The Brazilian Institute of Geography and Statistics and Author.
}
\usage{
perfume_prices_cpi_br
}
\description{
This dataset contains biweekly CPI data (IPCA and IPCA-15) for perfumes in Brazil and perfume prices collected from websites (30-day average vs. previous 30-day average).
}
\examples{
data(perfume_prices_cpi_br)
head(perfume_prices_cpi_br)
}
\keyword{datasets}
