% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{phillips_br}
\alias{phillips_br}
\title{Brazilian Phillips Curve Variables}
\format{
A data frame with 76 rows and 5 variables
}
\source{
The Brazilian Central Bank and Author's calculations.
}
\usage{
phillips_br
}
\description{
This dataset contains data for estimating a Phillips curve for Brazil.
}
\examples{
data(phillips_br)
head(phillips_br)
}
\keyword{datasets}
