% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{retail_sales_br}
\alias{retail_sales_br}
\title{Brazilian Retail Sales Survey (PMC)}
\format{
A data frame with 552 rows and 3 variables
}
\source{
The Brazilian Institute of Geography and Statistics.
}
\usage{
retail_sales_br
}
\description{
This dataset contains the monthly overall retail sales index for Brazil (PMC), both seasonally and non-seasonally adjusted.
}
\examples{
data(retail_sales_br)
head(retail_sales_br)
}
\keyword{datasets}
